+++ Tests/LiveKitCoreTests/DataTrack/DataTrackTests.swift
@@ -200,30 +200,6 @@ struct DataTrackTests {
         }
     }
 
-    // MARK: - Many Tracks
-
-    @Test
-    func publishManyTracks() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-        ]) { rooms in
-            let room = rooms[0]
-            let count = 64 // Conservative vs Rust's 256 — faster CI.
-
-            var tracks: [LocalDataTrack] = []
-            for i in 0 ..< count {
-                let track = try await room.localParticipant.publishDataTrack(name: "track-\(i)")
-                tracks.append(track)
-            }
-
-            #expect(tracks.count == count)
-            for (i, track) in tracks.enumerated() {
-                #expect(track.info.name == "track-\(i)")
-                #expect(track.isPublished)
-            }
-        }
-    }
-
     // MARK: - Send AsyncSequence
 
     @Test
@@ -370,261 +346,6 @@ struct DataTrackTests {
             #expect(try await pubRecorder.waitFor(.participantLocalUnpublish) == localSid)
             #expect(try await subRecorder.waitFor(.roomRemoteUnpublish) == remoteSid)
             #expect(try await subRecorder.waitFor(.participantRemoteUnpublish) == remoteSid)
-        }
-    }
-
-    // MARK: - Join-Time Tracks
-
-    /// A track published before a participant joins surfaces via the JoinResponse.
-    @Test
-    func receivesTrackPublishedBeforeJoin() async throws {
-        let roomName = UUID().uuidString
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(roomName: roomName, canPublishData: true),
-        ]) { pubRooms in
-            _ = try await pubRooms[0].localParticipant.publishDataTrack(name: "pre-join")
-
-            // The subscriber joins *after* the publish; its watcher is the delegate from creation,
-            // so it catches the publish delivered during connect (via the JoinResponse). A distinct
-            // identity avoids colliding with the publisher in the same room.
-            let watcher = DataTrackWatcher(expectedName: "pre-join")
-            try await TestEnvironment.withRooms([
-                RoomTestingOptions(delegate: watcher, roomName: roomName, identity: "subscriber", canSubscribe: true),
-            ]) { _ in
-                let remoteTrack = try await watcher.waitForTrack()
-                #expect(remoteTrack.info.name == "pre-join")
-            }
-        }
-    }
-
-    // MARK: - Publication Lifetime
-
-    /// The SDK retains the publication, so dropping the returned handle does not unpublish it
-    /// (matching JS; the raw Rust handle unpublishes on drop).
-    @Test
-    func retainsPublicationWhenHandleDropped() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-            RoomTestingOptions(canSubscribe: true),
-        ]) { rooms in
-            let subscriber = rooms[1]
-            let watcher = DataTrackWatcher(expectedName: "retained")
-            subscriber.delegates.add(delegate: watcher)
-
-            // Publish and immediately discard the handle.
-            _ = try await rooms[0].localParticipant.publishDataTrack(name: "retained")
-
-            let remoteTrack = try await watcher.waitForTrack()
-            #expect(remoteTrack.info.name == "retained")
-            // Still published a beat later — the dropped handle did not tear it down.
-            try await Task.sleep(nanoseconds: 1_000_000_000)
-            #expect(remoteTrack.isPublished)
-        }
-    }
-
-    // MARK: - Reconnect
-
-    /// A published data track survives the publisher's full reconnect: the session-scoped manager
-    /// republishes it under a new SID, and the subscriber's existing ``RemoteDataTrack`` carries
-    /// over — its SID is reassigned in place, with no unpublish/republish events fired.
-    @Test
-    func trackSurvivesPublisherFullReconnect() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-            RoomTestingOptions(canSubscribe: true),
-        ]) { rooms in
-            let publisher = rooms[0]
-            let subscriber = rooms[1]
-
-            _ = try await publisher.localParticipant.publishDataTrack(name: "survives-reconnect")
-            // Confirm the subscriber sees the initial publication.
-            let remoteTrack = try await subscriber.waitForDataTrack(name: "survives-reconnect")
-            let originalSid = remoteTrack.info.sid
-
-            // No unpublish/republish events should fire on the subscriber during the reconnect.
-            let recorder = DataTrackDelegateRecorder()
-            subscriber.delegates.add(delegate: recorder)
-            try await publisher.startReconnect(reason: .debug, nextReconnectMode: .full)
-
-            // The existing track object survives; its SID rotates once the track is republished.
-            let deadline = Date().addingTimeInterval(15)
-            while remoteTrack.info.sid == originalSid, Date() < deadline {
-                try await Task.sleep(nanoseconds: 200_000_000)
-            }
-            let newSid = remoteTrack.info.sid
-            #expect(newSid != originalSid)
-            #expect(remoteTrack.info.name == "survives-reconnect")
-
-            // The participant's track map follows the SID reassignment.
-            let participant = try #require(subscriber.remoteParticipants.values.first)
-            #expect(participant.dataTracks[newSid] === remoteTrack)
-            #expect(participant.dataTracks[originalSid] == nil)
-
-            // Continuity, not re-announcement: no events fired on the subscriber.
-            await #expect(throws: LiveKitError.self) {
-                _ = try await recorder.waitFor(.roomRemotePublish, timeout: 2)
-            }
-            await #expect(throws: LiveKitError.self) {
-                _ = try await recorder.waitFor(.roomRemoteUnpublish, timeout: 2)
-            }
-        }
-    }
-
-    /// Frames keep flowing across a quick reconnect: `SyncState.publishDataTracks` preserves the
-    /// publication and the resumed transports keep the subscription, so the same stream delivers.
-    @Test
-    func dataTrackSurvivesQuickReconnect() async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-            RoomTestingOptions(canSubscribe: true),
-        ]) { rooms in
-            let publisher = rooms[0]
-            let subscriber = rooms[1]
-
-            let watcher = DataTrackWatcher(expectedName: "sync-state")
-            subscriber.delegates.add(delegate: watcher)
-            let track = try await publisher.localParticipant.publishDataTrack(name: "sync-state")
-            let stream = try await watcher.waitForTrack().subscribe()
-
-            // Quick reconnect (resume). nextReconnectMode: .quick keeps the first attempt on the
-            // resume path, which sends SyncState (incl. publishDataTracks) and preserves transports.
-            try await publisher.startReconnect(reason: .debug, nextReconnectMode: .quick)
-
-            // Push in the background; assert at least one post-reconnect frame arrives (bounded, so
-            // a broken publication fails cleanly instead of hanging).
-            let payload = Data("after-quick-reconnect".utf8)
-            let pusher = Task {
-                while !Task.isCancelled {
-                    try? track.tryPush(frame: .now(payload: payload))
-                    try? await Task.sleep(nanoseconds: 100_000_000)
-                }
-            }
-            defer { pusher.cancel() }
-
-            let received = await withTaskGroup(of: Bool.self) { group in
-                group.addTask {
-                    for await frame in stream.values where frame.payload == payload {
-                        return true
-                    }
-                    return false
-                }
-                group.addTask {
-                    try? await Task.sleep(nanoseconds: 15_000_000_000)
-                    return false
-                }
-                let first = await group.next() ?? false
-                group.cancelAll()
-                return first
-            }
-            #expect(received, "Frames should keep flowing after a quick reconnect")
-        }
-    }
-
-    // MARK: - Concurrency
-
-    /// A multi-track concurrent-push stress scenario.
-    struct PushScenario: CustomTestStringConvertible {
-        let name: String
-        let trackCount: Int
-        let framesPerTrack: Int
-        let payloadSize: Int
-        var testDescription: String { name }
-
-        /// Many small frames across several tracks.
-        static let manySmall = PushScenario(name: "manySmall", trackCount: 8, framesPerTrack: 32, payloadSize: 8)
-        /// Fewer large multi-packet frames across several tracks.
-        static let largeFrames = PushScenario(name: "largeFrames", trackCount: 4, framesPerTrack: 8, payloadSize: 64 * 1024)
-    }
-
-    /// A received frame: the stream it arrived on, plus its tagged (track, sequence) header.
-    struct ReceivedFrame {
-        let stream: Int
-        let track: UInt32
-        let seq: UInt32
-    }
-
-    /// Builds a frame payload tagged with (trackIndex, sequence), padded to `size` bytes.
-    private static func makePayload(track: Int, seq: Int, size: Int) -> Data {
-        var trackIndex32 = UInt32(track)
-        var seq32 = UInt32(seq)
-        var data = Data(bytes: &trackIndex32, count: 4)
-        data.append(Data(bytes: &seq32, count: 4))
-        if size > 8 { data.append(Data(count: size - 8)) }
-        return data
-    }
-
-    /// Reads the tagged (track, sequence) header from a payload received on `stream`.
-    private static func parseFrame(_ payload: Data, stream: Int) -> ReceivedFrame {
-        let track = payload.withUnsafeBytes { $0.loadUnaligned(fromByteOffset: 0, as: UInt32.self) }
-        let seq = payload.withUnsafeBytes { $0.loadUnaligned(fromByteOffset: 4, as: UInt32.self) }
-        return ReceivedFrame(stream: stream, track: track, seq: seq)
-    }
-
-    /// Publishes several tracks, then pushes every frame on every track at once. Exercises the
-    /// Rust manager's per-track send queues and the bridge's packet dispatch under contention.
-    /// Each frame is tagged with (trackIndex, sequence); the unreliable channel may drop frames,
-    /// but whatever arrives must reach the right track with no duplicates or corruption.
-    @Test(arguments: [PushScenario.manySmall, .largeFrames])
-    func concurrentPush(_ scenario: PushScenario) async throws {
-        try await TestEnvironment.withRooms([
-            RoomTestingOptions(canPublishData: true),
-            RoomTestingOptions(canSubscribe: true),
-        ]) { rooms in
-            let framesPerTrack = scenario.framesPerTrack
-
-            // Publish each track and subscribe to it on the remote side.
-            var locals: [LocalDataTrack] = []
-            var streams: [DataTrackStream] = []
-            for i in 0 ..< scenario.trackCount {
-                let watcher = DataTrackWatcher(expectedName: "multi-\(i)")
-                rooms[1].delegates.add(delegate: watcher)
-                try await locals.append(rooms[0].localParticipant.publishDataTrack(name: "multi-\(i)"))
-                try await streams.append(watcher.waitForTrack().subscribe())
-            }
-
-            // All received frames (flat), drained concurrently with the burst.
-            let received = StateSync<[ReceivedFrame]>([])
-            var consumers: [Task<Void, Never>] = []
-            for (idx, stream) in streams.enumerated() {
-                consumers.append(Task {
-                    for await frame in stream.values {
-                        guard frame.payload.count >= 8 else { continue }
-                        let parsed = Self.parseFrame(frame.payload, stream: idx)
-                        received.mutate { $0.append(parsed) }
-                    }
-                })
-            }
-
-            // Push every frame on every track concurrently.
-            await withTaskGroup(of: Void.self) { group in
-                for (trackIndex, track) in locals.enumerated() {
-                    for seq in 0 ..< framesPerTrack {
-                        let payload = Self.makePayload(track: trackIndex, seq: seq, size: scenario.payloadSize)
-                        // try? — a full send queue under the burst is expected, not a failure.
-                        group.addTask { try? track.tryPush(frame: DataTrackFrame(payload: payload)) }
-                    }
-                }
-                await group.waitForAll()
-            }
-
-            // Let in-flight frames settle, then stop consuming.
-            try await Task.sleep(nanoseconds: 3_000_000_000)
-            for consumer in consumers {
-                consumer.cancel()
-            }
-
-            let all = received.copy()
-            #expect(!all.isEmpty, "Expected some frames to arrive")
-            for streamIndex in 0 ..< scenario.trackCount {
-                let frames = all.filter { $0.stream == streamIndex }
-                let routedCorrectly = frames.allSatisfy { $0.track == UInt32(streamIndex) }
-                let seqs = frames.map(\.seq)
-                let noDuplicates = Set(seqs).count == seqs.count
-                let inRange = seqs.allSatisfy { $0 < UInt32(framesPerTrack) }
-                #expect(routedCorrectly, "Stream \(streamIndex) received a frame from another track")
-                #expect(noDuplicates, "Stream \(streamIndex) received a duplicate frame")
-                #expect(inRange, "Stream \(streamIndex) received a corrupted sequence")
-            }
         }
     }
 }
